controller commit

+++ Speedcam/src/main/java/com/example/Speedcam/Speedcam/service/impl/FineServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.Speedcam.Speedcam.service.impl;
 import com.example.Speedcam.Speedcam.entity.Fine;
 import com.example.Speedcam.Speedcam.entity.mapperDTO.FineDTO;
 import com.example.Speedcam.Speedcam.entity.mapperDTO.MapperDTO;
+import com.example.Speedcam.Speedcam.exception.NotFoundException;
 import com.example.Speedcam.Speedcam.repository.FineRepository;
 import com.example.Speedcam.Speedcam.service.FineService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,18 +32,19 @@ public class FineServiceImpl implements FineService {
     }
 
     @Override
-    public Fine findFineById(Integer id) {
+    public Fine findFineById(Integer id) throws NotFoundException {
         Optional<Fine> f = fineRepository.findById(id);
         if(f.isEmpty()){
+            throw new NotFoundException("Fine with incorrect id");
         }
         return f.get();
     }
 
     @Override
-    public void updateFine(Integer id, Fine f) {
+    public void update(Integer id, Fine f) throws NotFoundException {
         Fine fine = findFineById(id);
         if(fine==null){
-
+            throw new NotFoundException("Fine with incorrect id");
         }
         fine.setDate(f.getDate());
         fine.setDeadline(f.getDeadline());
@@ -50,8 +52,11 @@ public class FineServiceImpl implements FineService {
     }
 
     @Override
-    public void deleteFine(Integer id) {
+    public void delete(Integer id) throws NotFoundException {
         Fine f = findFineById(id);
+        if(f==null){
+            throw new NotFoundException("Fine with incorrect id");
+        }
         fineRepository.delete(f);
     }
 }
