+++ Speedcam/src/main/java/com/example/Speedcam/Speedcam/service/impl/RegistryServiceImpl.java
@@ -1,10 +1,9 @@
 package com.example.Speedcam.Speedcam.service.impl;
 
-import com.example.Speedcam.Speedcam.entity.Fine;
 import com.example.Speedcam.Speedcam.entity.Registry;
-import com.example.Speedcam.Speedcam.entity.mapperDTO.FineDTO;
 import com.example.Speedcam.Speedcam.entity.mapperDTO.MapperDTO;
 import com.example.Speedcam.Speedcam.entity.mapperDTO.RegistryDTO;
+import com.example.Speedcam.Speedcam.exception.NotFoundException;
 import com.example.Speedcam.Speedcam.repository.RegistryRepository;
 import com.example.Speedcam.Speedcam.service.RegistryService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,18 +34,19 @@ public class RegistryServiceImpl implements RegistryService {
     }
 
     @Override
-    public Registry findRegistryById(Integer id) {
+    public Registry findRegistryById(Integer id)throws NotFoundException {
         Optional<Registry> r = registryRepository.findById(id);
         if(r.isEmpty()){
+            throw new NotFoundException("Registry not found");
         }
         return r.get();
     }
 
     @Override
-    public void updateRegistry(Integer id, Registry r) {
+    public void update(Integer id, Registry r) throws NotFoundException {
         Registry registry = findRegistryById(id);
         if(registry==null){
-
+            throw new NotFoundException("Registry not found");
         }
         registry.setName(r.getName());
         registry.setSurname(r.getSurname());
@@ -55,8 +55,11 @@ public class RegistryServiceImpl implements RegistryService {
     }
 
     @Override
-    public void deleteRegistry(Integer id) {
+    public void delete(Integer id) throws NotFoundException {
         Registry r = findRegistryById(id);
+        if(r==null){
+            throw new NotFoundException("Registry not found");
+        }
         registryRepository.delete(r);
     }
 }
